move download button to executions tab
